Move `CardTerminal` to teleport
