+++ web/packages/teleport/src/Login/CardTerminal/CardTerminal.jsx
