+++ web/packages/teleport/src/Login/CardTerminal/index.js
